Create docker-entrypoint.sh
This is a shellscript used to configure the container at runtime as
opposed to building the configuration in. It can be upgraded later with
tasks that would be good to run every time on the start of a container,
such as file ownership in docker volumes or database migrations.

For now, whenever the container is run with the /app/lemmy command (the
default), the script writes a /config/config.hjson file with values
pulled from environment variables.