Remove unnecessary API POST responses
Many POST endpoints return unnecessary data, such as a comment like
returning a full comment view including comment, creator, post,
community and more. All of this is likely unnecessary because the
client already fetched it in order to show the like button. Instead
a simple `success: true` is returned.

This is still work in progress, there are many more that can be removed.