Ensure new posts in NSFW communities are marked NSFW (#5649)
Clients exposing the NSFW setting to users will typically do so by providing a
checkbox, which is either checked or unchecked, but there is no third option
to let the server decide. As a result, virtually all posts created through
user interfaces will have a bool value for NSFW, so this logic likely is used
only for posts created by scripts/bots.

In #5310 the behavior for new posts was changed to default the NSFW flag to
the community's NSFW setting, but it only did so as a default if no value was
provided when creating a post via API.

This changes the logic to force all new posts created via API or received from
federation to be NSFW if the community they're posted in is marked NSFW,
otherwise use the value provided by the creator, with a default of false if
nothing is provided.

It also ensures that this logic is used for filtering out NSFW posts on
instances blocking NSFW content.