Improve media deletion logic
Media replacements for avatars, banners, and icons now only delete previously
used media if it was uploaded by the person performing the change or if that
person is an admin.

Removing avatars, banners, and icons has previously incorrectly retained them,
this has also been addressed by using the same deletion logic for both
replacements with new media and also removal of media in those places.

Updating community metadata now validates that the user has appropriate
permissions before executing media replacement code to delete old media.

Media deletions via API now validate that the user performing the deletion
either owns the upload or is an admin. This is a partial backport of #5317 and
#5588 without the associated DB migrations and the requirement of having a
correct pict-rs API key set up in Lemmy's configuration. The media deletion
token provided via API is now completely ignored and Lemmy only evaluates media
ownership or admin status to decide whether deletion is allowed.

For community banner and icon updates this no longer deletes old media unless
it's been uploaded by the current user or the person changing it is an admin.
Cleanup of old media in these cases will return in 1.0, which has a different
approach to dealing with this type of media.