Fall back to original url if thumbnail generation fails
I regret seeing failed thumbnails no longer linking to their original images anymore but instead showing broken images on the front-end. With the current situation people are unable to see large images from other sites.

I would like to suggest this workaround, to give the original image url as the link for now, the way it used to be. 

When an error is detected, it simply returns the original url of the image.